requirements.txt
# Core audio processing
librosa>=0.10.0
soundfile>=0.12.0
praat-parselmouth>=0.4.3
tgt>=1.4.4

# ML / Deep Learning
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Classical ML baseline
xgboost>=2.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Visualization & analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Config
pyyaml>=6.0
